Time: 92 ms (71.86%), Space: 198.5 MB (92.87%) - LeetHub

+++ 3453-separate-squares-i/3453-separate-squares-i.cpp
@@ -47,8 +47,8 @@ public:
                 break;
             }
 
-            long ia = static_cast<long long>((ansTop) * 1e5);
-            long ib = static_cast<long long>((ansBottom+1e-6) * 1e5);
+            long ia = static_cast<long>((ansTop) * 1e5);
+            long ib = static_cast<long>((ansBottom+1e-6) * 1e5);
             // cout << "ia:" << ia << ", ib:" << ib << endl;
             if (ia == ib) {
                 break;
